ProyectoDjango/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.12.4
fonttools==4.39.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.10.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.5